requirements.txt
annotated-doc==0.0.4
anyio==4.12.1
av==17.0.0
beautifulsoup4==4.14.3
certifi==2026.2.25
chardet==7.2.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
crcmod==1.7
ctranslate2==4.7.1
decorator==5.2.1
Deprecated==1.3.1
exceptiongroup==1.3.1
faster-whisper==1.2.1
ffmpeg-python==0.2.0
filelock==3.25.2
flatbuffers==25.12.19
fsspec==2026.2.0
future==1.0.0
google==3.0.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.2
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.9.1.4
nvidia-cudnn-cu12==9.20.0.48
onnxruntime==1.23.2
opencc-python-reimplemented==0.1.7
packaging==26.0
pillow==12.0.0
protobuf==7.34.1
py==1.11.0
pycryptodome==3.23.0
Pygments==2.19.2
pyreadline3==3.5.4
pysrt==1.1.2
pysubs2==1.8.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
retry==0.9.2
rich==14.3.3
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
sympy==1.13.1
tokenizers==0.22.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tos==2.9.0
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
volcengine==1.0.219
volcengine-python-sdk==5.0.21
wrapt==1.16.0
yt-dlp==2026.3.17